Modularise artist info routes
